Add a fusesoc flow to dv/formal for generating the fileset
This lets fusesoc do the heavy lifting in identify the correct files for us.
Fusesoc is already extensively used for this purpose for synthesis and simulation.

This also allows us to automatically patch the fusesoc-provided src files to work
around some current restrictions in the formal flow, without modifying the main
repository tree.

Signed-off-by: Harry Callahan <hcallahan@lowrisc.org>