[dv/formal] Remove -fun2wires 2:rX
Previously we used -fun2wires 2:rX so that we could track which registers
were being used. This was leading to combinational loops, and so had
already been removed from CHERIoT-Ibex. We can still 'track' them using
Ibex's own interpretation of which registers are relevant, and marking
the rest as x.