[pmp] MSB and LSB of PMP address now params
Instead of hardcoding 33 and 2 all over the code base as the most
significant and least significant bit of the PMP address, this change
creates appropriate parameters in the Ibex package that can be re-used
in the rest of the code base.

This commit also replaces some hard-coded 16 values with PMP_MAX_REGIONS
which was already a parameter in the Ibex package.

I thought of adding a parameter for the least significant bit of the PMP
address while analysing a question posed by Shutong Jin during the
HACK@CHES'25 event.