[CSR] Clarify PMP config assignment
Functionally pmp_mseccfg and pmp_mseccfg_q are the same when PMP is
enabled. However, the RTL is a bit more clear if both the MML and RLB
register are referenced through the pmp_mseccfg_q register name.

I thought of this clarification while analysing a question posed by
Shutong Jin during the HACK@CHES'25 event.