[dv] Fix the `riscv_dret_test`
The test checked for the exception handler to launch and then waited for
the core to flush its pipeline. However, the flush happened immediately
after the exception, which led the test to miss the first exception
handler and the `check_priv_mode` to fail, since the check only executed
after the flush leaving the exception handlers. This also means that on
later executions, we were not checking the exception handlers as we go,
but instead, the exception handler from the exception before.
Accordingly, this test only failed if Ibex initially started in USER
mode. Otherwise, it was anyways always in MACHINE mode