[rtl] Drop empty ifdef pair
This has been hanging around for a while: some git bisection finds
that it was gutted with commit c03cc91 (in 2021).

I'm also dropping the "Register File" heading. Thinking about it the
register file *does* contain the CSRs (so it's sort of reasonable to
have here) but there's no obvious hierarchy in the way the comments
are laid out. (Historically, I'm pretty certain this is just a line
that got left around when other lines moved)